app/requirements.txt
sqlalchemy
fastapi
uvicorn
psycopg2-binary
python-dotenv
scikit-learn
pandas
numpy